Remove need for container window
Previously the FramelessWindowHint option was used to get a borderless
window. This had the drawback that these types of windows cannot be
minimized. As a workaround, we gave the counter window a parent
container window. By minimizing the parent container, the counter would
be minimized too. This had issues like

- how to hide the parent window
- how to prevent focus from going to the parent window

This commit switches to a new method of hiding the window frame via
masks.